testing/web-platform/tests/DOMEvents/tests/submissions/Microsoft/converted/Determined.candidate.EventListeners.html

<!DOCTYPE html>
<html>
<head>
<title> Dispatch additional events inside an event listener </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>

<table id="table" border="1" style="display: none">
    <tbody id="table-body">
    <tr id="table-row">
        <td id="table-cell">Shady Grove</td>
        <td>Aeolian</td>
    </tr>
    <tr id="parent">
        <td id="target">Over the river, Charlie</td>
        <td>Dorian</td>
    </tr>
    </tbody>
</table>

<script>

    var EVENT = "bar";
    var TARGET = document.getElementById("target");
    var PARENT = document.getElementById("parent");
    var TBODY = document.getElementById("table-body");
    var TABLE = document.getElementById("table");
    var BODY = document.body;
    var HTML = document.documentElement;
    var CurrentTargets = [window, document, HTML, BODY, TABLE, TBODY, PARENT, TARGET];
    var ExpectResult = [window, document, HTML, BODY, TABLE, TBODY, PARENT, TARGET,
                        TARGET, PARENT, TBODY, TABLE,  BODY, HTML, document, window];
    var ActualResult = [];
    var ExpectListeners = [0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1];
    var ActualListeners = [];

    var description = "Test Description: " +
                      "Once determined, the current target's registered candidate event listeners cannot be changed; " +
                      "adding or removing listeners does not affect the current target's candidate event listeners.";

    var EventHandlers = [ function(evt){ TestEvent(evt, 0); },
                          function(evt){ TestEvent(evt, 1); },
                          function(evt){ TestEvent(evt, 2); },
                          function(evt){ TestEvent(evt, 3); } ];

    test(function()
    {
        for (var i=0; i < CurrentTargets.length; i++)
        {
            CurrentTargets[i].addEventListener(EVENT, EventHandlers[0], true);
            CurrentTargets[i].addEventListener(EVENT, EventHandlers[1], false);
        }

        var evt = document.createEvent("Event");
        evt.initEvent(EVENT, true, true);
        TARGET.dispatchEvent(evt);

        assert_array_equals(ActualResult, ExpectResult, "ActualResult");
        assert_array_equals(ActualListeners, ExpectListeners, "ActualListeners");

    }, description);

    function TestEvent(evt, i)
    {
        ActualResult.push(evt.currentTarget);
        ActualListeners.push(i);

        if ((evt.BUBBLING_PHASE != evt.eventPhase) && (1 != evt.currentTarget.foo))
        {
            evt.currentTarget.removeEventListener(EVENT, EventHandlers[0], true);
            evt.currentTarget.addEventListener(EVENT, EventHandlers[2], true);
            evt.currentTarget.foo = 1;
        }

        if ((evt.CAPTURING_PHASE != evt.eventPhase) && (3 != evt.currentTarget.foo))
        {
            evt.currentTarget.removeEventListener(EVENT, EventHandlers[0], false);
            evt.currentTarget.addEventListener(EVENT, EventHandlers[3], false);
            evt.currentTarget.foo = 3;
        }
    }
</script>
</body>
</html>
